Issue #LR-323 feat: Course Progress Exhaust to capture number of attempts
